Move enum declaration
