+++ src/main/java/com/google/sps/utility/GmailUtility.java
@@ -21,7 +21,6 @@ import com.google.api.client.json.JsonFactory;
 import com.google.api.client.json.jackson2.JacksonFactory;
 import com.google.api.services.gmail.Gmail;
 import com.google.api.services.gmail.model.Message;
-import com.google.sps.model.MessageFormat;
 import java.io.IOException;
 import java.util.List;
 
@@ -31,6 +30,26 @@ import java.util.List;
  */
 public final class GmailUtility {
   private GmailUtility() {}
+
+  /**
+   * Encapsulates possible values for the "format" query parameter in the Gmail GET message method
+   * FULL: Returns full email message data METADATA: Returns only email message ID, labels, and
+   * email headers MINIMAL: Returns only email message ID and labels; does not return the email
+   * headers, body, or payload. RAW: Returns the full email message data with body content in the
+   * raw field as a base64url encoded string;
+   */
+  public enum MessageFormat {
+    FULL("full"),
+    METADATA("metadata"),
+    MINIMAL("minimal"),
+    RAW("raw");
+
+    public final String formatValue;
+
+    MessageFormat(String formatValue) {
+      this.formatValue = formatValue;
+    }
+  }
 
   /**
    * Creates a query string for Gmail. Use in search to return emails that fit certain restrictions
